+++ tests/api_explorers/snapshots/snap_test_api_explorers.py
@@ -279,6 +279,67 @@ snapshots['test_playground_explorer_produces_html 1'] = '''<!DOCTYPE html>
   <script>window.addEventListener('load', function (event) {
       GraphQLPlayground.init(document.getElementById('root'), {
         // options as 'endpoint' belong here
+        
+      })
+    })</script>
+</body>
+</html>
+'''
+
+snapshots['test_playground_explorer_produces_html_with_settings 1'] = '''<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="user-scalable=no, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, minimal-ui">
+  <title>Hello world!</title>
+  <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/graphql-playground-react/build/static/css/index.css" />
+  <link rel="shortcut icon" href="//cdn.jsdelivr.net/npm/graphql-playground-react/build/favicon.png" />
+  <script src="//cdn.jsdelivr.net/npm/graphql-playground-react/build/static/js/middleware.js"></script>
+</head>
+
+<body>
+  <div id="root">
+    <style>
+      body {
+        background-color: rgb(23, 42, 58);
+        font-family: Open Sans, sans-serif;
+        height: 90vh;
+      }
+
+      #root {
+        height: 100%;
+        width: 100%;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+      }
+
+      .loading {
+        font-size: 32px;
+        font-weight: 200;
+        color: rgba(255, 255, 255, .6);
+        margin-left: 20px;
+      }
+
+      img {
+        width: 78px;
+        height: 78px;
+      }
+
+      .title {
+        font-weight: 400;
+      }
+    </style>
+    <img src='//cdn.jsdelivr.net/npm/graphql-playground-react/build/logo.png' alt=''>
+    <div class="loading"> Loading
+      <span class="title">Hello world!</span>
+    </div>
+  </div>
+  <script>window.addEventListener('load', function (event) {
+      GraphQLPlayground.init(document.getElementById('root'), {
+        // options as 'endpoint' belong here
+        settings: {"editor.cursorShape": "block", "editor.fontFamily": "helvetica", "editor.fontSize": 24, "editor.reuseHeaders": true, "editor.theme": "light", "general.betaUpdates": false, "prettier.printWidth": 4, "prettier.tabWidth": 4, "prettier.useTabs": true, "request.credentials": "same-origin", "request.globalHeaders": {"hum": "test"}, "schema.polling.enable": true, "schema.polling.endpointFilter": "*domain*", "schema.polling.interval": 4200, "schema.disableComments": true, "tracing.hideTracingResponse": true, "tracing.tracingSupported": true, "queryPlan.hideQueryPlanResponse": true},
       })
     })</script>
 </body>